// barq-system-frontend/src/sdk/types.gen.ts
export type AdditionalManagedServicesOneEntity = {
  id: number;
  description: string;
  cta_label: string;
  logo_id: number;
  file_id: number;
  logo: MediaEntity;
  file: MediaEntity;
  additional_managed_services_one_id_additional_managed_services_one_translations: Array<AdditionalManagedServicesOneTranslationsEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type AdditionalManagedServicesOneTranslationsEntity = {
  id: number;
  additional_managed_services_one_id: number;
  additional_managed_services_one_translation_additional_managed_services_one?: AdditionalManagedServicesOneEntity | null;
  description: string;
  cta_label: string;
  language: 'ar' | 'en';
};

export type AdditionalManagedServicesTwoEntity = {
  id: number;
  description: string;
  logo_id: number;
  images: number;
  logo: MediaEntity;
  additional_managed_services_two_images: Array<MediaEntity>;
  additional_managed_services_two_id_additional_managed_services_two_translations: Array<AdditionalManagedServicesTwoTranslationsEntity>;
  additional_managed_services_two_id_managed_service_download_form: Array<ManagedServiceDownloadFormEntity>;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type AdditionalManagedServicesTwoTranslationsEntity = {
  id: number;
  additional_managed_services_two_id: number;
  additional_managed_services_two_translation_additional_managed_services_two?: AdditionalManagedServicesTwoEntity | null;
  description: string;
  language: 'ar' | 'en';
};
// ...
export type type = 'soc' | 'cybersecurity' | 'grc';
// ...
export type ManagedCybersecurityServicesDetailsEntity = {
  id: number;
  description: string;
  cta_button_text: string;
  logo_id: number;
  logo: MediaEntity;
  managed_cybersecurity_services_details_id_managed_cybersecurity_services_details_translations: Array<ManagedCybersecurityServicesDetailsTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type ManagedCybersecurityServicesDetailsTranslationsEntity = {
  id: number;
  managed_cybersecurity_services_details_id: number;
  managed_cybersecurity_services_details_translation_managed_cybersecurity_services_details?: ManagedCybersecurityServicesDetailsEntity | null;
  description: string;
  cta_button_text: string;
  language: 'ar' | 'en';
};

export type ManagedGrcServicesDetailsEntity = {
  id: number;
  description: string;
  title: string;
  managed_grc_services_details_id_managed_grc_services_details_translations: Array<ManagedGrcServicesDetailsTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type ManagedGrcServicesDetailsTranslationsEntity = {
  id: number;
  managed_grc_services_details_id: number;
  managed_grc_services_details_translation_managed_grc_services_details?: ManagedGrcServicesDetailsEntity | null;
  description: string;
  title: string;
  language: 'ar' | 'en';
};

export type ManagedServiceCardsEntity = {
  id: number;
  description: string;
  bullet_one: string;
  bullet_two: string;
  type: string;
  image_id: number;
  logo_id: number;
  image: MediaEntity;
  logo: MediaEntity;
  managed_service_cards_id_managed_service_cards_translations: Array<ManagedServiceCardsTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type ManagedServiceCardsTranslationsEntity = {
  id: number;
  managed_service_cards_id: number;
  managed_service_cards_translation_managed_service_cards?: ManagedServiceCardsEntity | null;
  description: string;
  bullet_one: string;
  bullet_two: string;
  language: 'ar' | 'en';
};

export type ManagedServiceDownloadFormEntity = {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  phone_number: string;
  phone_number_key: string;
  position: string;
  managed_soc_services_details_id: number;
  managed_soc_services_details: ManagedSocServicesDetailsEntity;
  additional_managed_services_two_id: number;
  additional_managed_services_two: AdditionalManagedServicesTwoEntity;
  created_at: string;
  updated_at: string;
  deleted_at: string;
};

export type ManagedServiceHeroEntity = {
  id: number;
  title: string;
  sub_title: string;
  image_id: number;
  image: MediaEntity;
  logos: Array<MediaEntity>;
  managed_service_hero_id_managed_service_hero_translations: Array<ManagedServiceHeroTranslationsEntity>;
  created_at: string;
  updated_at: string;
};

export type ManagedServiceHeroTranslationsEntity = {
  id: number;
  managed_service_hero_id: number;
  managed_service_hero_translation_managed_service_hero?: ManagedServiceHeroEntity | null;
  title: string;
  sub_title: string;
  language: 'ar' | 'en';
};

export type ManagedSocServicesDetailsEntity = {
  id: number;
  description: string;
  cta_button_text: string;
  logo_id: number;
  file_id: number;
  logo: MediaEntity;
  file: MediaEntity;
  managed_soc_services_details_id_managed_soc_services_details_translations: Array<ManagedSocServicesDetailsTranslationsEntity>;
  managed_soc_services_details_id_managed_service_download_form: Array<ManagedServiceDownloadFormEntity>;
  created_at: string;
  updated_at: string;
};

export type ManagedSocServicesDetailsTranslationsEntity = {
  id: number;
  managed_soc_services_details_id: number;
  managed_soc_services_details_translation_managed_soc_services_details?: ManagedSocServicesDetailsEntity | null;
  description: string;
  cta_button_text: string;
  language: 'ar' | 'en';
};
// ...
export type id = 'ASC' | 'DESC';

export type title = 'ASC' | 'DESC';

export type description = 'ASC' | 'DESC';

export type image_id = 'ASC' | 'DESC';

export type created_at = 'ASC' | 'DESC';

export type updated_at = 'ASC' | 'DESC';
// ...
export type sub_title = 'ASC' | 'DESC';
// ...
export type logo_id = 'asc' | 'desc';
// ...
export type cta_label = 'asc' | 'desc';

export type cta_label_action2 = 'asc' | 'desc';

export type form_sub_title = 'asc' | 'desc';

export type file_id = 'asc' | 'desc';
// ...
export type deleted_at = 'asc' | 'desc';
// ...
export type number = 'asc' | 'desc';
// ...
export type cta_button_text = 'asc' | 'desc';
// ...
export type bullet_one = 'asc' | 'desc';

export type bullet_two = 'asc' | 'desc';
// ...
export type ManagedServiceHeroControllerReadResponse = {
  data?: Array<ManagedServiceHeroEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type ManagedServiceCardsControllerFindAllResponse = {
  data?: Array<ManagedServiceCardsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type ManagedSocServicesDetailsControllerReadResponse = {
  data?: Array<ManagedSocServicesDetailsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type ManagedCybersecurityServicesDetailsControllerReadResponse = {
  data?: Array<ManagedCybersecurityServicesDetailsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type ManagedGrcServicesDetailsControllerFindAllResponse = {
  data?: Array<ManagedGrcServicesDetailsEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AdditionalManagedServicesOneControllerReadResponse = {
  data?: Array<AdditionalManagedServicesOneEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};
// ...
export type AdditionalManagedServicesTwoControllerReadResponse = {
  data?: Array<AdditionalManagedServicesTwoEntity>;
  meta?: {
    total?: number;
    skip?: number;
    take?: number;
  };
};